+++ modules/generator/overrides.go
@@ -3,6 +3,7 @@ package generator
 import (
 	"github.com/grafana/tempo/modules/generator/registry"
 	"github.com/grafana/tempo/modules/overrides"
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
 type metricsGeneratorOverrides interface {
@@ -14,6 +15,7 @@ type metricsGeneratorOverrides interface {
 	MetricsGeneratorProcessorSpanMetricsHistogramBuckets(userID string) []float64
 	MetricsGeneratorProcessorSpanMetricsDimensions(userID string) []string
 	MetricsGeneratorProcessorSpanMetricsIntrinsicDimensions(userID string) map[string]bool
+	MetricsGeneratorProcessorSpanMetricsFilterPolicies(userID string) []filterconfig.FilterPolicy
 }
 
 var _ metricsGeneratorOverrides = (*overrides.Overrides)(nil)